Split process policy contracts from boundary tests
